cmd/trace-agent: join service_mapper* and trace_service_extractor* into service*.go
